requirements.txt
transformers==4.43.2
bitsandbytes==0.43.2
accelerate==0.33.0
fastapi==0.103.2
uvicorn==0.23.0
pydantic==2.8.1
motor==3.2.0
requests==2.31.0
python-dotenv==1.0.0
transformers==4.43.2
bitsandbytes==0.43.2
accelerate==0.33.0
torch==2.0.1
pexpect==4.8.0
deepface==0.0.79
opencv-python==4.8.0.76
numpy==1.24.3
aiohttp==3.8.5
whisper==1.1.10